requirements.txt
django
django-import-export
django-filter
django-queryset-csv
gunicorn